#12 Moved SrResources into separate files
