+++ ts_src/services/speedrun/index.ts
@@ -1,0 +1,7 @@
+export * from './category.resource';
+export * from './game.resource';
+export * from './leaderboard.resource';
+export * from './level.resource';
+export * from './links.resource';
+export * from './run.resource';
+export * from './variable.resource';
