+++ ts_src/services/speedrun/links.resource.ts
@@ -1,0 +1,4 @@
+export interface Links {
+    rel?:                    string;
+    uri:                    string;
+}
